+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Release 2021-11-30
 
   * Implement phpGettext class for better locale handling on Windows.
   * Enable vote editing.
+  * Add "alphabetically" as an option to sort answers when displayed.
 
 ## v0.1.1
 Release 2021-04-06